requirements.txt
# MetroPulse Baku — Python ETL dependencies
# Install: pip install -r requirements.txt

psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.30
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
